// generated/src/app/api-models/cap-report-work.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapReportWork {
  seqCcalcId: number;
  seqCapRptWork: number;
  seqCapRptDetailWork: number;
  capModelId: string;
  capMonth: Date;
  vendorId: string;
  vendorFullName: string;
  providerId: string;
  provLastName: string;
  provFirstName: string;
  provMiddleInitial: string;
  providerType: string;
  specialtyType: string;
  groupId: string;
  parentGroupId: string;
  groupShortName: string;
  subscriberId: string;
  personNumber: string;
  memberLastName: string;
  memberFirstName: string;
  memberMiddleInitial: string;
  memberAge: number;
  memberGender: string;
  planCode: string;
  copaymentAmt: number;
  capPool1Amt: number;
  capPool1Withhold: number;
  capPool2Amt: number;
  capPool2Withhold: number;
}